+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Calendar Versioning](https://calver.org).
 ### Added
 - [TUI] Added display of cURL commands for resource requests for UPDATE and DELETE operations.
 - [TUI] Added support for sending CREATE and UPDATE requests in the text UI.
+- [TUI] Added copy of CSV data to the clipboard for instance resources.
 
 
 ### Experimental